requirements.txt
PyPDF4
natsort
pillow